git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@4932 01821946-851b-0410-b7f4-f1477c70e57a

+++ render/fastrender.cpp
@@ -85,7 +85,7 @@ if (faceP->nFrame == gameData.app.nMineRenderCount)
 if (faceP - FACES.faces >= gameData.segs.nFaces)
 	return 0;
 #endif
-short	i, j, nKey = faceP->nBaseTex;
+int	i, j, nKey = faceP->nBaseTex;
 
 if (nKey < 0)
 	return 0;
